# openapi.yml
openapi: 3.0.0
info:
  title: Shop API
  version: 1.0.0
  description: JWT 기반 쇼핑몰 API

servers:
  - url: http://localhost:3000
    description: Development server

paths:
  /auth/signup:
    post:
      summary: 회원가입
      description: 이메일과 비밀번호로 새로운 사용자를 등록합니다.
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: 사용자 이메일
                  example: user@example.com
                password:
                  type: string
                  format: password
                  description: 사용자 비밀번호 (최소 6자 이상 권장)
                  example: password123
      responses:
        '201':
          description: 회원가입 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    description: 사용자 ID
                    example: 1
                  email:
                    type: string
                    description: 사용자 이메일
                    example: user@example.com
                  nickname:
                    type: string
                    description: 자동 생성된 닉네임
                    example: user
                  message:
                    type: string
                    description: 성공 메시지
                    example: User registered successfully
        '400':
          description: 잘못된 요청 (필수 필드 누락, 이미 가입된 이메일 등)
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: email and password are required
        '409':
          description: 이미 가입된 이메일
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 409
                  message:
                    type: string
                    example: User with this email already exists

  /auth/login:
    post:
      summary: 로그인
      description: 이메일과 비밀번호로 사용자를 인증하고 JWT 토큰을 발급합니다.
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: 사용자 이메일
                  example: user@example.com
                password:
                  type: string
                  format: password
                  description: 사용자 비밀번호
                  example: password123
      responses:
        '200':
          description: 로그인 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    description: JWT Access Token (1시간 유효)
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  refreshToken:
                    type: string
                    description: JWT Refresh Token (7일 유효)
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  user:
                    type: object
                    properties:
                      id:
                        type: number
                        example: 1
                      email:
                        type: string
                        example: user@example.com
                      nickname:
                        type: string
                        example: user
        '400':
          description: 잘못된 요청 (필수 필드 누락)
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: email and password are required
        '401':
          description: 인증 실패 (잘못된 이메일 또는 비밀번호)
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 401
                  message:
                    type: string
                    example: Invalid email or password

  /users:
    get:
      summary: 모든 사용자 조회
      description: 등록된 모든 사용자의 목록을 조회합니다.
      tags:
        - Users
      responses:
        '200':
          description: 사용자 목록 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: 사용자 생성
      description: 새로운 사용자를 생성합니다.
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - nickname
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  example: password123
                nickname:
                  type: string
                  example: John Doe
      responses:
        '201':
          description: 사용자 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: 잘못된 요청

  /users/{id}:
    get:
      summary: 특정 사용자 조회
      description: JWT 토큰이 필요합니다. 사용자 ID로 특정 사용자의 정보를 조회합니다.
      tags:
        - Users
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
          description: 사용자 ID
          example: 1
      responses:
        '200':
          description: 사용자 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: 인증되지 않음 (JWT 토큰 필요)
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 401
                  message:
                    type: string
                    example: No token provided
        '404':
          description: 사용자를 찾을 수 없음
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 404
                  message:
                    type: string
                    example: User with id 1 not found

  /products:
    get:
      summary: 상품 목록 조회
      description: 카테고리 코드(outer, top, bottom)로 필터링 가능한 상품 목록을 조회합니다.
      tags:
        - Products
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
            enum: [outer, top, bottom]
          description: 카테고리 코드
      responses:
        '200':
          description: 상품 목록 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '400':
          description: 잘못된 카테고리
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: invalid category

  /products/{id}:
    get:
      summary: 상품 상세 조회
      description: 상품 ID로 상품 상세와 옵션(사이즈별 재고)을 조회합니다.
      tags:
        - Products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
          description: 상품 ID
          example: 1
      responses:
        '200':
          description: 상품 상세 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: 잘못된 상품 ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: valid product id is required
        '404':
          description: 상품을 찾을 수 없음
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 404
                  message:
                    type: string
                    example: Product with id 1 not found

  /products/{id}/cart:
    post:
      summary: 상품 장바구니 추가
      description: 로그인 사용자 장바구니에 상품 옵션(사이즈)을 추가합니다.
      tags:
        - Products
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
          description: 상품 ID
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - size
              properties:
                size:
                  type: string
                  enum: [s, m, l, xl, xxl]
                  example: m
                quantity:
                  type: number
                  example: 1
      responses:
        '201':
          description: 장바구니 추가 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartItem'
        '400':
          description: 요청 검증 실패(사이즈/수량/재고)
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: requested quantity exceeds stock
        '401':
          description: 인증되지 않음
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 401
                  message:
                    type: string
                    example: No token provided

  /cart-items/{id}/quantity:
    patch:
      summary: 장바구니 수량 변경
      description: 로그인 사용자의 장바구니 항목 수량을 지정한 값으로 변경합니다.
      tags:
        - Cart
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
          description: 장바구니 아이템 ID
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - quantity
              properties:
                quantity:
                  type: number
                  example: 2
      responses:
        '200':
          description: 수량 변경 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartItem'
        '400':
          description: 수량 검증 실패 또는 재고 초과
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: requested quantity exceeds stock
        '401':
          description: 인증되지 않음
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 401
                  message:
                    type: string
                    example: No token provided
        '404':
          description: 장바구니 항목을 찾을 수 없음
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 404
                  message:
                    type: string
                    example: Cart item with id 1 not found

  /cart-items:
    get:
      summary: 장바구니 조회
      description: 로그인 사용자의 장바구니 항목 목록을 조회합니다.
      tags:
        - Cart
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 장바구니 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CartItem'
        '401':
          description: 인증되지 않음
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 401
                  message:
                    type: string
                    example: No token provided

  /cart-items/{id}:
    delete:
      summary: 장바구니 항목 삭제
      description: 로그인 사용자의 장바구니 항목을 삭제합니다.
      tags:
        - Cart
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
          description: 장바구니 아이템 ID
          example: 1
      responses:
        '200':
          description: 삭제 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 1
                  message:
                    type: string
                    example: Cart item deleted successfully
        '401':
          description: 인증되지 않음
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 401
                  message:
                    type: string
                    example: No token provided
        '404':
          description: 장바구니 항목을 찾을 수 없음
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 404
                  message:
                    type: string
                    example: Cart item with id 1 not found

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: number
          description: 사용자 ID
          example: 1
        email:
          type: string
          format: email
          description: 사용자 이메일
          example: user@example.com
        password:
          type: string
          description: 해시된 비밀번호
          example: $2b$10$...
        nickname:
          type: string
          description: 사용자 닉네임
          example: John Doe
        isActive:
          type: boolean
          description: 사용자 활성 상태
          example: true
        createAt:
          type: string
          format: date-time
          description: 생성 날짜
          example: 2024-03-16T11:59:21.625Z
        updateAt:
          type: string
          format: date-time
          description: 수정 날짜
          example: 2024-03-16T11:59:21.625Z

    ProductCategory:
      type: object
      properties:
        id:
          type: number
          example: 1
        code:
          type: string
          example: outer
        name:
          type: string
          example: 아우터
        createAt:
          type: string
          format: date-time
        updateAt:
          type: string
          format: date-time

    ProductOption:
      type: object
      properties:
        id:
          type: number
          example: 1
        productId:
          type: number
          example: 1
        size:
          type: string
          enum: [s, m, l, xl, xxl]
          example: m
        stock:
          type: number
          example: 10
        createAt:
          type: string
          format: date-time
        updateAt:
          type: string
          format: date-time

    Product:
      type: object
      properties:
        id:
          type: number
          example: 1
        name:
          type: string
          example: Sample Outer Jacket
        categoryId:
          type: number
          example: 1
        description:
          type: string
          example: Sample product description.
        material:
          type: string
          example: Cotton 100%
        origin:
          type: string
          example: Korea
        washingMethod:
          type: string
          example: Hand wash separately
        price:
          type: number
          example: 99000
        salePrice:
          type: number
          nullable: true
          example: 79000
        category:
          $ref: '#/components/schemas/ProductCategory'
        options:
          type: array
          items:
            $ref: '#/components/schemas/ProductOption'
        createAt:
          type: string
          format: date-time
        updateAt:
          type: string
          format: date-time

    CartItem:
      type: object
      properties:
        id:
          type: number
          example: 1
        userId:
          type: number
          example: 1
        productOptionId:
          type: number
          example: 2
        quantity:
          type: number
          example: 1
        createAt:
          type: string
          format: date-time
        updateAt:
          type: string
          format: date-time

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT 토큰 인증

security:
  - BearerAuth: []